Move TaskGraph to file
