+++ src/sciline/pipeline.py
@@ -18,7 +18,9 @@ from typing import (
     overload,
 )
 
-from .scheduler import DaskScheduler, NaiveScheduler, Scheduler
+from sciline.task_graph import TaskGraph
+
+from .scheduler import Graph, Scheduler
 
 T = TypeVar('T')
 
@@ -37,10 +39,6 @@ class AmbiguousProvider(Exception):
 
 Provider = Callable[..., Any]
 Key = type
-Graph = Dict[
-    Key,
-    Tuple[Callable[..., Any], Dict[str, Key]],
-]
 
 
 def _is_compatible_type_tuple(
@@ -227,39 +225,3 @@ class Pipeline:
         else:
             graph = self.build(keys)
         return TaskGraph(graph=graph, keys=keys, scheduler=scheduler)
-
-
-class TaskGraph:
-    def __init__(
-        self,
-        *,
-        graph: Graph,
-        keys: type | Tuple[type, ...],
-        scheduler: Optional[Scheduler] = None,
-    ) -> None:
-        self._graph = graph
-        self._keys = keys
-        if scheduler is None:
-            try:
-                scheduler = DaskScheduler()
-            except ImportError:
-                scheduler = NaiveScheduler()
-        self._scheduler = scheduler or DaskScheduler()
-
-    def compute(self, keys: Optional[type | Tuple[type, ...]] = None) -> Any:
-        """
-        Compute the result of the graph.
-
-        Parameters
-        ----------
-        keys:
-            Optional list of keys to compute. This can be used to override the keys
-            stored in the graph instance. Note that the keys must be present in the
-            graph as intermediate results, otherwise KeyError is raised.
-        """
-        if keys is None:
-            keys = self._keys
-        if isinstance(keys, tuple):
-            return self._scheduler.get(self._graph, list(keys))
-        else:
-            return self._scheduler.get(self._graph, [keys])[0]
